requirements.txt
flask
Flask-Compress
python-dotenv
pinecone-client
git+https://github.com/dglazkov/polymath.git